codegen(sdk/go): reject deprecated pragmas on required args
Here, we consider "optional" any arg which, if removed, would break the
function call.

It's a bit more than optional, because we could have variadic args,
default and optional args. All of those lead to a non breakage of the
function. Deprecating those would be fine

We fail fast in module_funcs.go, when a required argument carries a +deprecated pragma

Signed-off-by: Guillaume de Rouville <guillaume@dagger.io>