chore(sdk/python): do not return runtime files
After a `dagger init --sdk=python` or a `dagger develop` on a
python module, do not return the `sdk/python/runtime` files as part
of the generated code.
Those files are not necessary for the python module itself, they are
only necessary to build the python SDK on the engine side.

This helps to keep the user module code cleaner and more understandable.

Signed-off-by: Yves Brissaud <yves@dagger.io>