handle per-LLB-vertex deduplication by retrying in case of stale session
The buildkit solver still merges operations if the LLB vertex digest is
the exact same. This fixes the code such that if we end up with a stale
session, the operation will be retried rather than giving all clients
waiting on the result an error just because one client disconnected.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>