fix: always gracefully shut down the listening
When creating a listener on the host, it was possible that we could
leave a dangling connection.

If somehow we failed to dial the connection, or we failed to read or
write info from the connection, we just abandoned the connection and
left it in an open state. This seems to be an oversight.

I encountered this originally when I misconfigured a service without a
healthcheck and tried to connect to it immediately. The container hadn't
started in time, so the dial failed, but incoming connection stayed open
indefinitely.

Signed-off-by: Justin Chadwell <me@jedevc.com>