llm: run destructive calls before readonly calls
I've observed LLMs tend to run batches of tool calls that involve 1. a
change to make, and 2. a test to run assuming the change is made. Right
now when that happens we guarantee that the test runs WITHOUT the
changes, which confuses the LLM.

Instead we should do all the changes and _then_ the read-only calls,
which will read the freshest state.

Signed-off-by: Alex Suraci <alex@dagger.io>