add filesync telemetry for root paths (#11315)
* add filesync telemetry for root paths

adds OTel spans for filesync uploads. This providers better visibility
into what's effectively being uploaded to the context. It only reports
root paths given that in most cases this will otherwise produce a very
large amount of spans which will put a lot of pressure on the tracing
backend.

We're also collecting the amount of written bytes as well as a the total duration for each root path

Signed-off-by: Marcos Lilljedahl <marcosnils@gmail.com>