chore(sdk/python): do not pull uv image if not required
With the bundled python SDK, `uv` and `uvx` are packaged inside the
engine. In that case we don't need to pull an image to get those tools.

Only pull the image if it's required. On my machine this saves almost
1.5s.

If `uv` and `uvx` are not found in the `dist` (bundled) directory, pull
the image as usual.

The benefit is really on the first call. Otherwise the image would be
cached. But the first call is still very important as it's creates a
first impression.

This changes also mount `uv` and `uvx` instead of copying the directory
with filters. This is a very small improvement, but still good to have.

Signed-off-by: Yves Brissaud <yves@dagger.io>