feat: add a "debug" boolean option inside the dagger.json's SDK field
In order to improve and formalize jumping into the context of the
runtime, this commit adds an option to the dagger.json (easiest option)
to jump into the container runtime.

It's a good DX loop to work on the auth:

1. Just add a `debug: true` field -->

```json
{
  "name": "dagger-dev",
  "engineVersion": "v0.19.4",
  "sdk": {
    "source": "go",
    "debug": true
  }
}
```

Important note: As we now lazy load the runtime, this leads to annoying cache
hits (not being able to jump into the container).

In order to avoid that, when this option is set, we add random to the moduleSource's digest to burst the moduleSource cache

Signed-off-by: Guillaume de Rouville <guillaume@dagger.io>