fix: replace main.py in tests
Python modules are using two files:
- __init__.py importing the following
- main.py containing the module code

The way test code was working was kind of a hack: the main.py
was still there and the code was put into __init__.py and removing
the import. That way, when loaded, the main.py file was ignored.
As we are moving to a static parser, all *.py files will be parsed
and imported.
So it's better to write the code into main.py, keep the import line
in __init__.py as it shouldn't have an impact.

Signed-off-by: Yves Brissaud <yves@dagger.io>