Revert "core: re-add content hashing of git and dir module sources (#11292)"
This reverts commit 21411b1c53d385a3b56852f2c9dd7cf727343749.

This content hashing actually creates severe issues for private git
repos when there's an overlap in content; e.g. multiple git repos with
different auth that actually have the same content.

We should revisit this but it's not worth the trouble in our CI right
now (where we have lots of git repos with different auth that have the
exact same commits).

Signed-off-by: Erik Sipsma <erik@sipsma.dev>