fix loading cached contextual dirs (#11350)
* fix overwhelmingly large error strings from ids

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

* use _contextDirectory API to fix loading of cached results

By loading contextual directory args with a dedicated API, we can manage
to force reloading of them from the correct context even if the
contextual dir shows up in a cached result of a function call from an
older session.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

* add integ test

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

* handle contextual files too

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

* fix git module source case

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

* update telemetry test

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

* add changelog

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

* Revert "core: re-add content hashing of git and dir module sources (#11292)"

This reverts commit 21411b1c53d385a3b56852f2c9dd7cf727343749.

This content hashing actually creates severe issues for private git
repos when there's an overlap in content; e.g. multiple git repos with
different auth that actually have the same content.

We should revisit this but it's not worth the trouble in our CI right
now (where we have lots of git repos with different auth that have the
exact same commits).

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

---------

Signed-off-by: Erik Sipsma <erik@sipsma.dev>