telemetry: bump sqlite write timeout
We've seen in CI some SQLITE BUSY errors popping up again in the
telemetry client dbs. It's *possible* (but far from confirmed) that
other performance improvements may be resulting in us writing to those
dbs faster and thus making it easier to hit that error.

Trying out a bump of the timeout to see if it helps in the short term.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>