core/git: cache git.Remote() metadata in session
We currently run `git ls-remote` on every NewGitRepository, without any cache. This operation takes on average ~0.7s, and should be optimized a bit more.

This commit introduces a session-level caching on this step, meaning that multiple Remote() calls with a different dag.Git(ref, options) leading to the same ref to be cached.

Signed-off-by: Guillaume de Rouville <guillaume@dagger.io>