dagop: never use potentially stale client metadata (#11295)
* dagop: never use potentially stale client metadata

We see quite a few flakes around missing sessions in the midst of dagop.
It's always when integ tests are running in parallel and a lot that are
sharing a step will hit the same error.

I think it may be a timing thing where the client metadata stored in an
op is cached in dagql, then the unlazy happens later with timing such
that the stored client metadata is okay at first but becomes invalid
later due to the session it was from disconnecting.

Instead, I think we should just always use the client metadata of
whoever is invoking the operation.

This is much simpler as easy to think through now-a-days since we
deleted edge merging, no possible concerns with multiple clients mixing
data anymore.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

* handle per-LLB-vertex deduplication by retrying in case of stale session

The buildkit solver still merges operations if the LLB vertex digest is
the exact same. This fixes the code such that if we end up with a stale
session, the operation will be retried rather than giving all clients
waiting on the result an error just because one client disconnected.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

* add changelog

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

---------

Signed-off-by: Erik Sipsma <erik@sipsma.dev>