chore: remove engine.Tag and align dev publications
We are publishing dev versions based on `main`.
To align them with the tagged versions, keep only one single way
to identify the engine: Version as comming from the version module.

The version is defined as following:
- if this is a dev version: `v[major].[minor].[patch]-[timestamp]-dev-[digest]`
- if this is an untagged version: `v[major].[minor].[patch]-[timestamp]-[commit]`
- if this is a tagged version: `v[major].[minor].[patch]`

The `engine.Tag` (comming from Version.Tag) was also used to find
the right reference to pull non bundled SDKs like PHP, Java, Elixir.
The rule is:
- if this is a tagged version: use the tag
- otherwise, use the most recent tag existing on `main`

This is now available under `Version.TagOrCommit` and set to
`engine.SDKVersion`.

Signed-off-by: Yves Brissaud <yves@dagger.io>