fix python sdk tests to actually use sidecar dev engine (#11380)
* fix python sdk tests to actually use sidecar dev engine

The tests were being given a container with a dev engine attached via
service dep, but then were being run as nested execs, which overrode the
settings and made them not actually hit the dev engine they were
supposed to hit at all.

This adds and option to the tests to disable nested execs (which is
particular to the nested dev engine sidecar setup in our top-level
module) and uses that to run tests.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

* appease linter

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

---------

Signed-off-by: Erik Sipsma <erik@sipsma.dev>