ci: fix provision test on macos (#11388)
* ci: fix provision test on macos

The action douglascamata/setup-docker-macos-action doesn't work with the
new lima v2:
https://github.com/douglascamata/setup-docker-macos-action/issues/53

As the action picks by default the latest version of lima, go back to
the previous stable lima v1.2.2 that should be ok.

Signed-off-by: Yves Brissaud <yves@dagger.io>

* ignore errors on logs

Logs are great to understand better when something goes wrong. But
sometimes the thing that went wrong will also make this goes wrong.
So ignore errors when displaying logs to keep the real error more
visible.

Signed-off-by: Yves Brissaud <yves@dagger.io>

* run one call first, to warm up everything

Signed-off-by: Yves Brissaud <yves@dagger.io>

---------

Signed-off-by: Yves Brissaud <yves@dagger.io>