telemetry: publish metrics at most every 5s rather than 1s
We see *a lot* of logs about publishing metrics in the engine logs, I
believe due to the fact that it runs every 1s.

That 1s value was arbitrarily chosen when metrics were added. In
retrospect, it seems a bit excessive for everything (including all
function calls) to get metrics updated that frequently. 5s is still
close enough to "live" (imo) while also potentially reducing the burden
on the clientdbs a significant amount.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>