telemetry: skip some more noisy introspection related fields
Various function type fields show up a enormous number of times in
telemetry. These are the fields that show up more than 1000 times as
spans across a run of TestTelemetry:

1018 name="mkfile /schema.json"
1140 name=Container.withWorkdir
1154 name=Directory.withoutDirectory
1162 name=Container.withMountedDirectory
1254 name=TypeDef.withObject
1418 name=Container.withExec
1976 name="dagop.ctr Container.withExec"
1994 name=Function.withDescription
3038 name=Function.withArg
5538 name=Function.withCachePolicy
6364 name="POST /query"
6562 name=Query.function
6634 name=Function.withSourceMap
6638 name=TypeDef.withFunction
10310 name=Query.sourceMap

The overall total of all spans in the run was 84889.

After hiding some of the noisier and mostly uninteresting fields,
the total spans across the run was down to 48264.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>