core: handle nil returns in various places (#11420)
A user reported a panic from ContainerDagOp where it was hitting either
a nil return from LoadType or a return that when unwrapped is nil.

I don't have a repro for that yet, so the code there is just an update
to turn that into an error and include more debug information to help
track it down if/when it happens again.

However while attempting repros of it I managed to hit independent
panics involving modules returning various permutations of nil things in
lists. Fixed those here and included integ test coverage.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>