engine: handle stale sessions in dag-op filesync too
Previously, we added a fix for stale sessions showing up in dag-op due
to per-LLB vertex deduplication in Server.clientFromIDs.

There's another codepath possibly impacted by the same problem when
trying to get a session from the buildkit client. So we add the same fix
there too.

This theoretically may be the cause of some flakes we occasionally hit
with errors like:
```
failed to stat path: failed to get requester session: session for "XYZ" not found
```

Signed-off-by: Erik Sipsma <erik@sipsma.dev>