tui: prevent orphaned spans corrupting data
Something in our full test suite appears to be emitting orphaned spans.
They're all Docker/Containerd related, so maybe it's Docker running
somewhere in our suite?

When this happens, we've been promoting that orphaned span to the new
root span, and since we auto-cancel all running spans when the root span
is completed, that was leading to mass cancellations. Along the way
we'll track explicit state for debugging Canceled in that scenario.

Signed-off-by: Alex Suraci <alex@dagger.io>