core: fix service detaching race condition
When all clients of a service detach, we start a goroutine to stop the
service and only remove it from the Services tracker after the service
actually stops.

That gives rise to a race condition: if we detached, but then a client
came in an attached, it would attach to this "doomed" service, instead
of starting its own fresh instance.

Now we'll detect when a "running" service is actually stopping (bindings
== 0), wait for it to exit, and then start a fresh instance.

Signed-off-by: Alex Suraci <alex@dagger.io>