core/sdks: enable runtime debug mode (#11349)
* feat: add a "debug" boolean option inside the dagger.json's SDK field

In order to improve and formalize jumping into the context of the
runtime, this commit adds an option to the dagger.json (easiest option)
to jump into the container runtime.

It's a good DX loop to work on the auth:

1. Just add a `debug: true` field -->

```json
{
  "name": "dagger-dev",
  "engineVersion": "v0.19.4",
  "sdk": {
    "source": "go",
    "debug": true
  }
}
```

2. Then, call any dagger function (any return types will work). You will
   jump into the runtime container of that function (Go, Ts and Python
SDKs), prior the call

Important note: As we now lazy load the runtime, this leads to annoying cache
hits (not being able to jump into the container). In order to avoid
that, when this option is set, we burst the cache every time

Signed-off-by: Guillaume de Rouville <guillaume@dagger.io>

* feat(typescript): implement that option on the typescript SDK

Signed-off-by: Guillaume de Rouville <guillaume@dagger.io>

* feat(sdk/python): implement the logic on the Python SDK

Adds the terminal() option logic to the Python SDK to jump into the container when running any
dagger function with this debug option

In ModuleTypesExp, we now manually construct the runtime, as it was
being called in other places, leading to jumping inside the terminal
before we wanted to

Signed-off-by: Guillaume de Rouville <guillaume@dagger.io>

* fix: unify the step where we jump into the container

As pointed by Tom, all SDKs shall allow to jump into the runtime at the
same time (once everything is set).

This fixes my previous mistake where all the SDKs were not aligne

Signed-off-by: Guillaume de Rouville <guillaume@dagger.io>

---------

Signed-off-by: Guillaume de Rouville <guillaume@dagger.io>