sdk: use module's env
Currently when an SDK module uses a contextual dir it inherits the
context of the module that's using it, which doesn't seem like what
you'd want for an SDK module.

This affects Dang, which uses a contextual dir to bring in the Dang
language repo so the SDK can compile its entrypoint binary.

It looks like we already had this problem for our builtin SDKs, which
already use contextual dir args, but we actually handled it manually by
hardcoding an expected arg name and making the behavior optional.

Now it's consistent across built-in SDKs and module SDKs: the context is
always the SDK module's context.

Signed-off-by: Alex Suraci <alex@dagger.io>