tests: fix LocalCacheGC flake (#11501)
Some of the local cache pruning tests have been flaking for a bit in the
test cases where prunes are run manually.

Based on engine logs from passed/failed tests, it appeared that when the
prune call was made before the other client had its session fully
removed, the test would flake. This is because some parts of session
cleanup happen slightly async, so if the prune hit too early it couldn't
quite prune everything yet.

The fix here just adds retries similar to those in the other test cases.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>