engine: fix nested execs created by modules to not be module clients (#11458)
* engine: fix nested execs created by modules to not be module clients

There was some vestigial looking code when creating execs that resulted
in nested execs created by modules to be marked as being clients for
that module. This is subtly incorrect since they are not actually module
clients (i.e. clients for function calls of the module), they are just
execs created by the module.

This broke some cases of trying to load further modules from these
nested execs because filesync tries to find the "non-module parent
client", and would thus skip over the nested exec client up to the main
caller.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

* integ test

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

* fix service hostnames

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

---------

Signed-off-by: Erik Sipsma <erik@sipsma.dev>