engine: fix invalid overlay mount creation
We need to avoid the following type of race condition, which can result in invalid overlapping overlay mounts:
1. Engine creates random (unrelated to this exec) overlay mount like upperdir=B,lowerdir=A
2. We hit this code and start runc, which gets to the point where it has unshared its mount namespace but
   not yet pivot_root'd. In this state, since mount namespaces are forks of their parent, the overlay mount
   from (1) is visible in the runc processes mount namespace.
3. Engine unmounts the overlay from (1), but that does NOT unmount it from the runc process's mount namespace
4. Engine creates a new overlay mount like upperdir=C,lowerdir=B:C (i.e. upperdir from (1) is now lowerdir).
   This mount is invalid and technically hitting "undefined behavior" since B is still an upperdir in mounts
   that exist on the system.

We avoid this by starting the runc process in a clean mount namespace that was created during engine init before
any mounts existed, guaranteeing none are leaked into it. We setns to that clean mount namespace and then unshare
again to guarantee the namespace for the runc process is fully isolated. OpenTree+MoveMount are then used to
bind the mounts actually needed by the container into that isolated namespace so that runc can see them.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>