Reduce fsync overhead via overlay volatile option
Looked into the chown perf regression with ebpf and found that each chown was causing overlay to
call fsync on every file (within its copy-up internal functions), each of which took ~8ms on my
laptop. 8ms isn't a lot to a human but when you are serially chowning 100k files it adds up quickly.
Syscalls like this shouldn't be that slow generally, especially for the test case where each file is
only 1 byte.

It's not entirely clear to me why overlay makes this so slow. I can repro the slowness outside of
dagger with hand-crafted overlay mounts, but when I remove overlay from the equation and use the
same underlying filesystem it's much much faster.

This all reminded me of marcosnils's effort a long time ago to upstream support for the overlay
volatile option, which met a bunch of friction and didn't get merged. Thankfully, we no longer have
those problems!

`volatile` skips the fsyncs overlay does and fixes the perf problem for me. There's a good chunk of
trickiness and careful handling needed here since we sometimes need to remount overlays that were
previously mounted with the same upper/work dirs (cache mounts, exec-after-failure mounts, etc.), in
which case the leftover `<workdir>/work/incompat` dir needs to be removed to prevent overlay from
error'ing out (its attempt to warn us that the mount may have been corrupted after a previous crash).

Signed-off-by: Erik Sipsma <erik@sipsma.dev>