bug: fixes incorrect file caching (#11575)
This fixes a file caching bug, reported in #11552, which resulted from
an incorrect Digest() override, which ignored the file Source
(contents).

Removing the override will cause the digest to use the entire args
struct, which includes the Source field, which contains a digest of
the file being copied in, e.g. `Source:File@xxh3:97e0aa643fe2496b`

Signed-off-by: Alex Couture-Beil <alex@mofo.ca>