default to ignoring xattr copy failures
Some systems do not allow the engine to set SELinux xattrs when copying
a file. The engine ignores these in many places, but it's easy to miss.
This change updates the default behavior of copying to be ignoring these
errors so we don't have to squash it everywhere all the time.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>