chore: propagate the error instead of panicking (#11455)
qemuBins() panics when it fails to list binaries from the QEMU container image. When the registry is down or unreachable, this results in a crash with a stack trace instead of a clear error message.

This PR changes qemuBins() to return an error, consistent with how other registry fetches are handled in the same function (e.g., bin.Name() errors at lines 153-156).

Signed-off-by: Guillaume de Rouville <guillaume@dagger.io>